update remove and block functionality

+++ client/src/screens/message/ChatScreen.js
@@ -10,13 +10,15 @@ import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityI
 import FontAwesome from 'react-native-vector-icons/FontAwesome';
 import { io } from 'socket.io-client';
 import { message, auth } from '../../apis';
-import { SOCKET_URL } from '../../configs';
+import { ASSET_API_URL, SOCKET_URL } from '../../configs';
 import { useRoute, useNavigation } from '@react-navigation/native';
 import { useDispatch, useSelector } from 'react-redux';
 import { Icon, ListItem, Button, Text } from 'react-native-elements';
 import { stacks } from '../../constants/title';
 import { chatActions } from '../../redux/actions';
 import { Toast } from '../../helpers';
+import { HeaderBackButton } from '@react-navigation/elements';
+
 const ChatScreen = () => {
   // const socket = useRef();
   const [messages, setMessages] = useState([]);
@@ -32,16 +34,21 @@ const ChatScreen = () => {
   const [blocked, setBlocked] = useState(route.params.blocked);
   const dispatch = useDispatch();
   const socket = useSelector((state) => state.auth.socket);
-  console.log(blocked);
+
   const fetchMessages = async () => {
     try {
       const res = await message.getMessageByOtherUserId(receivedId, token);
-      return res.data.data;
+      return res.data;
     } catch (err) {
       if (err.response.status == 404) {
         return null;
       }
     }
+  };
+
+  const handleBack = () => {
+    dispatch(chatActions.updateCurrentChatRoom(null));
+    navigation.navigate('MessageStack');
   };
 
   useEffect(() => {
@@ -61,16 +68,19 @@ const ChatScreen = () => {
           }
         />
       ),
+      headerLeft: () => (
+        <HeaderBackButton style={{ marginLeft: 0 }} onPress={handleBack} />
+      ),
     });
   }, [route, navigation, chatId]);
 
   useEffect(() => {
     const initialize = async () => {
       const newMessages = await fetchMessages();
-      if (newMessages && newMessages.length > 0) {
+      if (newMessages && newMessages.data.length > 0) {
         setMessages(
-          newMessages.map((msg) => {
-            if (msg.isDeleted == false) {
+          newMessages.data.map((msg) => {
+            if (!msg.isDeleted) {
               return {
                 ...msg,
                 _id: msg._id,
@@ -97,13 +107,21 @@ const ChatScreen = () => {
             }
           }),
         );
-        setChatId(newMessages[0].chat);
-        dispatch(chatActions.updateSeenStatus(newMessages[0].chat));
+        setChatId(newMessages.data[0].chat);
+        dispatch(chatActions.updateSeenStatus(newMessages.data[0].chat));
+        setBlocked(newMessages.blockStatus.blocked);
+        setIsBlocked(newMessages.blockStatus.isBlocked);
       }
     };
     initialize();
     return () => {};
   }, []);
+
+  useEffect(() => {
+    if (chatId) {
+      dispatch(chatActions.updateCurrentChatRoom(chatId));
+    }
+  }, [chatId]);
 
   useEffect(() => {
     socket?.on('getMessage', (data) => {
@@ -115,6 +133,7 @@ const ChatScreen = () => {
           user: {
             _id: data.senderId,
           },
+          image: data.senderAvatar,
         };
 
         setMessages((previousMessages) =>
@@ -123,39 +142,24 @@ const ChatScreen = () => {
       }
     });
 
-    // socket?.on('blocked', (data) => {
-    //   if (user.id == data.userId){
-    //     set
-    //   }
-    // })
-  }, [socket]);
-  useEffect(() => {
-    socket?.on('beBlocked', (data) => {
-      if (user.id == data.receivedId) {
+    socket?.on('blockChat', (data) => {
+      if (user.id == data.userId && receivedId == data.receivedId) {
+        setBlocked(true);
+      } else if (user.id == data.receivedId && receivedId == data.userId) {
         setIsBlocked(true);
       }
     });
-  }, [socket]);
-  useEffect(() => {
-    socket?.on('blocked', (data) => {
-      if (user.id == data.userId) {
-        setBlocked(true);
-      }
-    });
-  }, [socket]);
-  useEffect(() => {
-    socket?.on('beUnblocked', (data) => {
-      if (user.id == data.receivedId) {
+
+    socket?.on('unblockChat', (data) => {
+      if (user.id == data.userId && receivedId == data.receivedId) {
+        setBlocked(false);
+      } else if (user.id == data.receivedId && receivedId == data.userId) {
         setIsBlocked(false);
-        console.log('vu hoang trung');
       }
     });
-  }, [socket]);
-  useEffect(() => {
+
     socket?.on('removeMess', (data) => {
-      console.log('data', data);
-      // console.log(messages[0]);
-      if (receivedId === data.userId) {
+      if (chatId && chatId == data.chatId && data.userId != user.id) {
         setMessages(
           messages.map((msg) => {
             if (msg._id == data._id) {
@@ -164,15 +168,14 @@ const ChatScreen = () => {
                 isDeleted: true,
                 text: 'Message unsent',
               };
-            } else {
-              return msg;
             }
+            return msg;
           }),
-          // previousMessages.filter((messages) => messages._id !== messageIdToDelete),
         );
       }
     });
-  }, [socket]);
+  }, [chatId, socket]);
+
   useEffect(() => {
     if (isLoadingEarlier) {
       handleLoadEarlier();
@@ -195,7 +198,7 @@ const ChatScreen = () => {
       const formattedEarlierMessages = earlierMessages.data.data.map((msg) => ({
         ...msg,
         _id: msg._id,
-        text: msg.content,
+        text: !msg.isDeleted ? msg.content : 'Message unsent',
         createdAt: msg.createdAt,
         user: {
           _id: msg.user._id,
@@ -220,22 +223,36 @@ const ChatScreen = () => {
           newMsgObj.text,
           token,
         );
+
+        const newMessage = sendResult.data.data;
+
         setMessages((previousMessages) =>
-          GiftedChat.append(previousMessages, messages),
+          GiftedChat.append(previousMessages, [
+            {
+              _id: newMessage._id,
+              text: newMessage.content,
+              createdAt: newMessage.createdAt,
+              isDeleted: newMessage.isDeleted,
+              user: {
+                _id: newMessage.user._id,
+                name: newMessage.user.username,
+              },
+            },
+          ]),
         );
 
-        const newMsg = sendResult.data.data;
-        dispatch(chatActions.updateSeenStatus(newMsg.chat._id));
+        dispatch(chatActions.updateSeenStatus(newMessage.chat._id));
 
         socket?.emit('sendMessage', {
-          chatId: newMsg.chat._id,
-          _id: newMsg._id,
+          chatId: newMessage.chat._id,
+          _id: newMessage._id,
           senderId: senderId,
+          senderAvatar: `${ASSET_API_URL}/${newMessage.user.avatar.fileName}`,
           receivedId: receivedId,
-          userName: receiverName,
-          userImg: receiverImg,
-          content: newMsgObj.text,
-          createdAt: newMsg.createdAt,
+          // userName: receiverName,
+          // userImg: receiverImg,
+          content: newMessage.content,
+          createdAt: newMessage.createdAt,
         });
       } catch (err) {
         console.log(err);
@@ -310,7 +327,13 @@ const ChatScreen = () => {
   };
 
   const onDelete = async (messageIdToDelete) => {
-    const deleteMess = await message.deleteMessage(messageIdToDelete, token);
+    try {
+      await message.deleteMessage(messageIdToDelete, token);
+    } catch (err) {
+      Toast.showFailureMessage('Error deleting message');
+      return;
+    }
+
     setMessages(
       messages.map((msg) => {
         if (msg._id == messageIdToDelete) {
@@ -324,33 +347,38 @@ const ChatScreen = () => {
       }),
     );
 
-    const messDelete = deleteMess.data.data;
     socket?.emit('deleteMessage', {
-      _id: messDelete._id,
+      _id: messageIdToDelete,
+      chatId,
       userId: user.id,
+      isLatest: messageIdToDelete == messages[0]._id,
     });
   };
 
   const onLongPress = (context, message) => {
-    const options = ['Copy', 'Delete Message', 'Cancel'];
-    const cancelButtonIndex = options.length - 1;
-    context.actionSheet().showActionSheetWithOptions(
-      {
-        options,
-        cancelButtonIndex,
-      },
-      (buttonIndex) => {
-        switch (buttonIndex) {
-          case 0:
-            Clipboard.setString(message.text);
-            break;
-          case 1:
-            console.log('delete');
-            onDelete(message._id);
-            break;
-        }
-      },
-    );
+    if (message.user._id == user.id && !message.isDeleted) {
+      console.log('message to delete: ', message);
+
+      const options = ['Copy', 'Delete Message', 'Cancel'];
+      const cancelButtonIndex = options.length - 1;
+      context.actionSheet().showActionSheetWithOptions(
+        {
+          options,
+          cancelButtonIndex,
+        },
+        (buttonIndex) => {
+          switch (buttonIndex) {
+            case 0:
+              Clipboard.setString(message.text);
+              break;
+            case 1:
+              console.log('delete');
+              onDelete(message._id);
+              break;
+          }
+        },
+      );
+    }
   };
   const onPressUnblockUser = async () => {
     try {
@@ -396,7 +424,7 @@ const ChatScreen = () => {
           <ListItem.Title>You have blocked this user</ListItem.Title>
           <Button
             onPress={onPressUnblockUser}
-            title="Unblock this user"
+            title="Unblock"
             color="#841584"
             accessibilityLabel="Learn more about this purple button"
           />
